// src/generated/server/worldmonitor/emergency/v1/service_server.ts
export interface RoadClosure {
  id: string;
  road: string;
  location: string;
  reason: string;
  status: string;
  island: string;
  latitude: number;
  longitude: number;
  detour: string;
  updatedAt: string;
}

export interface ListRoadClosuresRequest {
  island: string;
}

export interface ListRoadClosuresResponse {
  closures: RoadClosure[];
}